Fix deadlock due to side-effecting I/O writes
Problem:

We're currently using the Debug Module to read/write the processor's
memory.  Some I/O writes trigger reads and writes of the processor's
memory, which causes a dependency cycle through the processor, because
the processor and the Debug Module share the interface to memory.

Solution:

Change virtio so that it defers any processor memory access or other
actions until the I/O write has completed.

Also, because Linux is still using cached memory for DMA, this has the
advantage that we can halt the processor (which flushes the cache),
perform the memory operations, and then resume the processor, giving
the appearance of cache-coherent DMA.