Fixed deadlock when connecting dma_pcis
Problem:

When a 512-bit AXI fabric was interposed between the
AXI_Mem_Controller and the DDR pins, the memory system deadlocked
almost immediately.

AXI_Mem_Controller cannot be reset until it is ready, which occurs
after it fetches word 0 of memory. Because we were resetting the
fabric it was attempting to request data from, it never got a response
and never reached the ready state.

Solution:

All the modules reset on power on, so removed the calls to reset.