Fix VirtIO IRQ assignments
IRQ 0 is taken by the UART, so we should start the VirtIO IRQs at 1.

The PLIC has no ID 0, and so the source IDs it presents are 1 higher than the
external source IDs. The device tree gives the PLIC's source IDs, and so
the existing device tree was already wrong. With the new IRQ
assignments, the source IDs start at 2.

Also match the style of the rest of the device tree by using the clearer
interrupts-extended property.