First step toward loading ELF via DMA
AWSP2 now includes a 64KB BRAM for the "uncached" memory. It is
accessible from the P2 processor and from the host via the PCIS_DMA
port.

This adds support for loadELF to write to the BRAM and then run a
small memcpy helper on the P2 processor to copy from BRAM to cached
DRAM.

It is not enabled by default because it has not been tested on the
FPGA and is missing the code to mmap() the PCIS_DMA device node.