Separate out AWS SoC fabric from Flute internal fabric
This means the GFE Flute's internal fabric's ID fields can remain as
4-bit, and we extend them at the boundary between the core and the AWS
interconnect.

This is how it logically should be, as the two are independent, but
having this separation will also make it easier to support CHERI Flute
without needing lots of ifdefs everywhere to special-case it, instead
limiting the BlueStuff AXI ifdefs to just AWS_AXI4_Connection.bsv.