Copy AXI4 bug fixes from Flute
00312cc6b6 stopped even the GFE variant using Flute's internal fabric
implementation outside the core, but meant that we lost the bug fix for
when a fabric only has one master, as is the case for our memFabric.
Thus, sync the entirety of all our copies with the current versions in
the Flute submodule and pick up all the latest changes.