AWS_AXI4_Deburster.bsv: Fix major timing regression
Pulling in the updated copy from Flute introduced a major timing
regression. This implementation uses 64-bit % operators on the AR and AR
channels, which is extremely expensive and currently only hits about 23
MHz. Instead, we should be exploiting that wrapping bursts are only
powers of two in size (precisely so this can be done efficiently) and do
bitmasking instead.