Software changes for faster DMA-based VirtIO
This includes removing a bunch of now-unused memory access functions. We
do still need to defer queue notifications to another thread to ensure
we don't deadlock due to our own requests going through the same L1 as
the core. Unfortunately for now we also need to halt the core whilst
processing those notifications to avoid the aggressive PCIS timeout
biting us, but this could be avoided by using the XDMA block instead to
drive the DMA_PCIS interface, which has a 5s timeout in place of 8us. No
effort has been put into making the halt/resume clean as we really want
to be able to remove it, though it gives a useable workaround for now.