Move VirtIO process_io into its own thread and add stop method
Ideally this thread would do a blocking select on all the fd's, but the
virtio_console_can_write_data call, as well as the net device's
select_fill implementation, mean that the set of fds is conditional on
whether the virtqueues are ready to accept input, and there's no easy
way to wake up when those change. Instead, we use a 10ms timeout, which
is slightly wasteful, but means we should quickly react to such changes.