Use separate 100 MHz clock for the core and up SoC clock to 250 MHz
The AWS clock recipes only give you three A clock recipes, none of which
are useful for us given we can't meet timing at 125 MHz. Patching the
AWS scripts to use a different clock frequency also doesn't help, as
there are crossings between clock_xtra within the shell and
clock_main_a0 with no synchronizer, so running clock_main_a0 at 97 MHz
(fpga-load-local-image lets you have either 97 or 109 MHz, and we can't
meet timing at the latter) causes either setup or hold violations on
that crossing. Presumably if fpga-load-local-image is given a different
frequency, the input to that clock is scaled (otherwise it wouldn't
work), but it is unclear whether clock_main_a0 will be scaled in a
compounding manner (causing us to run even slower than 97 MHz) or just
overwritten with a 97 MHz clock. This would all lead to a rather hacky
solution that does not inspire confidence.

Instead, we should do what the docs point towards, namely using a
separate clock group with a recipe that gives you a real 100 MHz clock
out of the box. For this, we choose B5's clock_extra_b1. We are then
also able to change the A group's recipe to A1, which clocks the SoC and
shell at 250 MHz and should help to alleviate the overheads of the CDCs.
Unfortunately, the lack of a mkGSyncFIFO interacts poorly with AXI, so
we need additional Xactors with their own mkGFIFOFs to give the
axi_side's unguarded interfaces, but hopefully the faster shell will
compensate sufficiently.

The lack of a pessimistic notEmpty signal on the enq side of a sync FIFO
is also frustrating, and gives rise to rg_dmi_reads_queued. If we don't
care about reporting that in dmi_status then it can however go away.