loadelf.cpp: Turn into a proper ELF loader
ELF loaders are supposed to consult the program headers, not the section
headers, which are intended for linkers and debugging tools. This means
we are sure to load to the right physical addresses, and without relying
on the first header's LMA/VMA difference being the same as all the rest.
As part of this, also handle zeroing out the ends of segments like a
proper ELF loader should.

The only reason remaining to consult the section headers is for
detecting .htif/.tohost/.fromhost.