Support loading multiple ELF files
The entry point of the first will be used, as will the address for
.htif/.tohost/.fromhost if present. This is useful for using a BBL
configured using --without-payload so that the kernel can be provided
separately, avoiding the need to re-link BBL every time the kernel
changes, and is equivalent to OpenSBI's fw_jump.

Whilst here, remove the "using namespace std" bad practice.