Use raw mode for terminal input
This means control charcters are passed through directly to the
hardware, there is no line buffering, and no local echo. This provides
an experience similar to QEMU, and C-a is used as a prefix like QEMU in
order to send commands to the local instance (e.g. C-a x will quit the
host software in a controlled manner).

Since we no longer have ONLCR, we need to include the CR in our own
print statements. Some of these changes are unnecessary, but it does no
harm when ONLCR is set before we switch to raw mode, so we might as well
be consistent everywhere and not have to think about which bits of code
are reached in raw mode and which in cooked mode.