Add SiFive Test Finisher device
Unlike the HTIF, which requires modifying BBL's linker script to use an
I/O address, the SiFive Test Finisher has its address in the device tree
like a normal memory-mapped device, but otherwise its functionality is
very similar. BBL uses it as the primary means of powering off when
present, falling back on the HTIF (which doesn't work out of the box for
us) and then, finally, a WFI loop to spin.

Thus, provide an implementation of the SiFive Test Finisher so that BBL
is able to shut down completely. We only implement sifive,test0 and not
also sifive,test1; the latter adds an additional 0x7777 value for
resetting (which would not be much effort to refactor main.cpp to
support), but neither BBL nor OpenSBI provide an SBI call that lets you
reboot. Linux probably supports it if exposed via syscon-reboot, but
FreeBSD has no such driver (although it really should) and is wholly
reliant on SBI calls.

Whilst hooking this up, also make the HTIF riscv-tests device actually
exit and report the status.