Bump Flute_CHERI
We use mkCore_Synth to convert back to the old-style interface with
synthesisable AXI, as the SoC here is designed around that. It's a bit
silly as we then connect the interfaces to FIFOs for the clock domain
crossings (or for an implied deburster by the AXI4-to-AXI4-Lite
mkConnection for slave0), but fixing that is a bigger undertaking and
now isn't the time for that.