Use Clock Wizard IP rather than the more restrictive set of AWS recipes
Whilst 100 MHz for Flute is available from the set of recipes provided
by AWS, Toooba's 25 MHz is not. We could use mkClockDivider, but that
does not automatically get detected as a generated clock so requires
constraints that need to use internal Bluespec-generated names, and that
method of dividing a clock is not recommended by Xilinx regardless.
Thus, use the recommended IP instead.

Note that we are cascading MMCMs this way, and since we can't place ours
in a vertically-adjacent clock region to the one in the shell we have to
add a constraint telling Vivado this is ok.